meta-qti-audio: Add correct app type for unprocessed record
Regular capture usecases are configured with bit width as 24
instead of 16.
App type for unprocessed record is set to default value(69938) which
is used for primary input. Audio HAL picks the maximum bit width of
the inputs having same app type and configures bit width to 24
for all usecases using this app type.
Use dedicated app type for unprocessed record inputs which is already
defined and avoid conflict with primary input.

Change-Id: I63a92024aec7abf9b0810b11c3d910d409ea334d