audiohal and encoders depend on audiodlkm for exported headers
The encoders recipe requires that the audiodlkm must complete and the headers
are exported under work-shared/kernel-build-artifacts/. Hence, encoders
is made to depend on audiodlkm install task to ensure headers are exported.
Following error is resolved with this change:

 In file included from src/omx_aac_aenc.cpp:41:
 inc/omx_aac_aenc.h:61:10:
 fatal error: linux/msm_audio.h: No such file or directory
 #include <linux/msm_audio.h>

The audiohal compilation depends on headers exported from audiodlkm.
The audiohal configure task depends on audiodlkm install task to ensure
headers are exported to work-shared/kernel-build-artifacts/

 In file included from acdb.c:29:
 acdb.h:24:10:
 fatal error: linux/msm_audio_calibration.h: No such file or directory
 #include <linux/msm_audio_calibration.h>

 In file included from audio_extn/compress_in.c:41:
 ../hal/msm8916/platform.h:22:10:
 fatal error: sound/voice_params.h: No such file or directory
 #include <sound/voice_params.h>

Change-Id: I0f46a3af580f92a5889c1907ddfb87061629fad0