meta-qti-audio: fix audiohal HY11 build issue
audio_parsers.h is included by source code of audiohal, and Makefile of
audiohal trys to find headers inside the source code path, this is good
in a normal build. But in HY11 build, audio-parsers' header won't be
present in the source code path because it is qcom-proprietary and the
build fails.
Except for the source code path, audio-parsers's header will be installed
into audiohal.bb's recipe-sysroot dierectory. Thus, set the CFLAGS to that
path so the haeder can be found during HY11.

Change-Id: I6e489facfd06df827a9a073ad6a06c015f53dc47