meta-qti-audio: Modify ExecStart command for init_audio service
As busybox is replaced with toybox, current syntax of command
for ExecStart is not working and resulting in service start
failure.
echo is used with -n option to eliminate newline charecter
from the value which is being echoed.
Usage of backslash is removed to avoid wrong interpretation
of if condition.

Change-Id: I4063fcb67051cb217122c5d5b34a2028e1980f81