meta-qti-audio:do_compile add depend on securmsmdlkm do_compile
where two modules (audiodlkm and securitydlkm) are being built simultaneously, and one of them is causing a failure in the other. The failure happens during the kernel configuration step.
Incomplete Cleanup: One build might be leaving intermediate files in a state that causes the subsequent build to fail.
so, Serializing the Builds for mdm9607.

Change-Id: I48a16d99b930ad9413b49c7144471257c98ad3a1